delete method implemented

+++ application/src/main/java/com/finalgo/application/entity/Project.java
@@ -12,7 +12,7 @@ public class Project implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private Integer amount;
@@ -21,11 +21,11 @@ public class Project implements Serializable {
     @Column(name = "ownerUsername")
     private String ownerUsername;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -60,4 +60,6 @@ public class Project implements Serializable {
     public void setOwnerUsername(String ownerUsername) {
         this.ownerUsername = ownerUsername;
     }
+
+    
 }
